+++ packages/backend/src/git.ts
@@ -1,53 +1,72 @@
 import { CheckRepoActions, GitConfigScope, simpleGit, SimpleGitProgressEvent } from 'simple-git';
+import { mkdir } from 'node:fs/promises';
+import { env } from './env.js';
 
 type onProgressFn = (event: SimpleGitProgressEvent) => void;
 
-export const cloneRepository = async (cloneURL: string, path: string, onProgress?: onProgressFn) => {
-    const git = simpleGit({
-        progress: onProgress,
-    });
+export const cloneRepository = async (
+    remoteUrl: URL,
+    path: string,
+    onProgress?: onProgressFn
+) => {
     try {
-        await git.clone(
-            cloneURL,
-            path,
-            [
-                "--bare",
-            ]
-        );
+        await mkdir(path, { recursive: true });
 
-        await git.cwd({
+        const git = simpleGit({
+            progress: onProgress,
+        }).cwd({
             path,
-        }).addConfig("remote.origin.fetch", "+refs/heads/*:refs/heads/*");
+        })
+
+        await git.init(/*bare = */ true);
+
+        await git.fetch([
+            remoteUrl.toString(),
+            // See https://git-scm.com/book/en/v2/Git-Internals-The-Refspec
+            "+refs/heads/*:refs/heads/*",
+            "--progress",
+        ]);
     } catch (error: unknown) {
-        if (error instanceof Error) {
-            throw new Error(`Failed to clone repository: ${error.message}`);
+        const baseLog = `Failed to clone repository: ${path}`;
+
+        if (env.SOURCEBOT_LOG_LEVEL !== "debug") {
+            // Avoid printing the remote URL (that may contain credentials) to logs by default.
+            throw new Error(`${baseLog}. Set environment variable SOURCEBOT_LOG_LEVEL=debug to see the full error message.`);
+        } else if (error instanceof Error) {
+            throw new Error(`${baseLog}. Reason: ${error.message}`);
         } else {
-            throw new Error(`Failed to clone repository: ${error}`);
+            throw new Error(`${baseLog}. Error: ${error}`);
         }
     }
-}
+};
 
-
-export const fetchRepository = async (path: string, onProgress?: onProgressFn) => {
-    const git = simpleGit({
-        progress: onProgress,
-    });
-
+export const fetchRepository = async (
+    remoteUrl: URL,
+    path: string,
+    onProgress?: onProgressFn
+) => {
     try {
-        await git.cwd({
+        const git = simpleGit({
+            progress: onProgress,
+        }).cwd({
             path: path,
-        }).fetch(
-            "origin",
-            [
-                "--prune",
-                "--progress"
-            ]
-        );
+        })
+
+        await git.fetch([
+            remoteUrl.toString(),
+            "+refs/heads/*:refs/heads/*",
+            "--prune",
+            "--progress"
+        ]);
     } catch (error: unknown) {
-        if (error instanceof Error) {
-            throw new Error(`Failed to fetch repository ${path}: ${error.message}`);
+        const baseLog = `Failed to fetch repository: ${path}`;
+        if (env.SOURCEBOT_LOG_LEVEL !== "debug") {
+            // Avoid printing the remote URL (that may contain credentials) to logs by default.
+            throw new Error(`${baseLog}. Set environment variable SOURCEBOT_LOG_LEVEL=debug to see the full error message.`);
+        } else if (error instanceof Error) {
+            throw new Error(`${baseLog}. Reason: ${error.message}`);
         } else {
-            throw new Error(`Failed to fetch repository ${path}: ${error}`);
+            throw new Error(`${baseLog}. Error: ${error}`);
         }
     }
 }
@@ -72,6 +91,28 @@ export const upsertGitConfig = async (path: string, gitConfig: Record<string, st
             throw new Error(`Failed to set git config ${path}: ${error.message}`);
         } else {
             throw new Error(`Failed to set git config ${path}: ${error}`);
+        }
+    }
+}
+
+/**
+ * Unsets the specified keys in the git config for the repo at the given path.
+ * If a key is not set, this is a no-op.
+ */
+export const unsetGitConfig = async (path: string, keys: string[], onProgress?: onProgressFn) => {
+    const git = simpleGit({
+        progress: onProgress,
+    }).cwd(path);
+
+    try {
+        for (const key of keys) {
+            await git.raw(['config', '--unset', key]);
+        }
+    } catch (error: unknown) {
+        if (error instanceof Error) {
+            throw new Error(`Failed to unset git config ${path}: ${error.message}`);
+        } else {
+            throw new Error(`Failed to unset git config ${path}: ${error}`);
         }
     }
 }